a2a_friend_scheduling/nate_agent_crewai/pyproject.toml
[project]
name = "nate-scheduling-agent"
version = "0.1.0"
description = "Nate's scheduling agent for the A2A project."
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.5",
    "crewai[tools]>=0.126.0",
    "pydantic",
    "python-dotenv",
    "aioquic>=1.2.0",
    "cryptography>=42.0.0",
    "google-generativeai",
] 
